Cleaned up the repo for releasing the new updates
Revised documentation. Fixed the input country raster.
Fixed the taiwan and hong kong land rent disaggregation. Made the year-specific ref veg
thematic consistent with the valid land mask (now all regular outputs are on
the valid land mask). Replaced examples with new outputs.
Replaced ancillary data with new files. Ensured that scripts point to examples.
Renamed some files for consistency. Minor edits to some scripts for consistency and flexibility.
Made sure that everything works with the 18 AEZ outputs.
Deleted some unnecessary files. Made sure all diagnostics scripts work.

+++ src/read_aez_orig.c
@@ -90,7 +90,7 @@ int read_aez_orig(args_struct in_args, rinfo_struct *raster_info) {
 	}
 	
 	// read the data
-	num_read = fread(aez_bounds_orig, insize, ncells, fpin);
+	num_read = (int) fread(aez_bounds_orig, insize, ncells, fpin);
 	fclose(fpin);
 	if(num_read != ncells)
 	{
