+++ src/read_land_area_sage.c
@@ -92,7 +92,7 @@ int read_land_area_sage(args_struct in_args, rinfo_struct *raster_info) {
 	}
 	
 	// read the data
-	num_read = fread(land_area_sage, insize, ncells, fpin);
+	num_read = (int) fread(land_area_sage, insize, ncells, fpin);
 	fclose(fpin);
 	if(num_read != ncells)
 	{
