+++ src/read_soil_carbon.c
@@ -126,7 +126,7 @@ int read_soil_carbon(args_struct in_args, rinfo_struct *raster_info) {
     }
 
     // read the data and check for same size as the working grid
-    num_read = fread(wavg_array, insize, ncells, fpin);
+    num_read = (int) fread(wavg_array, insize, ncells, fpin);
     fclose(fpin);
     if(num_read != NUM_CELLS)
     {
@@ -153,7 +153,7 @@ int read_soil_carbon(args_struct in_args, rinfo_struct *raster_info) {
     }
 
     // read the data and check for same size as the working grid
-    num_read = fread(median_array, insize, ncells, fpin);
+    num_read = (int) fread(median_array, insize, ncells, fpin);
     fclose(fpin);
     if(num_read != NUM_CELLS)
     {
@@ -179,7 +179,7 @@ int read_soil_carbon(args_struct in_args, rinfo_struct *raster_info) {
     }
 
     // read the data and check for same size as the working grid
-    num_read = fread(min_array, insize, ncells, fpin);
+    num_read = (int) fread(min_array, insize, ncells, fpin);
     fclose(fpin);
     if(num_read != NUM_CELLS)
     {
@@ -205,7 +205,7 @@ int read_soil_carbon(args_struct in_args, rinfo_struct *raster_info) {
     }
 
     // read the data and check for same size as the working grid
-    num_read = fread(max_array, insize, ncells, fpin);
+    num_read = (int) fread(max_array, insize, ncells, fpin);
     fclose(fpin);
     if(num_read != NUM_CELLS)
     {
@@ -231,7 +231,7 @@ int read_soil_carbon(args_struct in_args, rinfo_struct *raster_info) {
     }
 
     // read the data and check for same size as the working grid
-    num_read = fread(q1_array, insize, ncells, fpin);
+    num_read = (int) fread(q1_array, insize, ncells, fpin);
     fclose(fpin);
     if(num_read != NUM_CELLS)
     {
@@ -257,7 +257,7 @@ int read_soil_carbon(args_struct in_args, rinfo_struct *raster_info) {
     }
 
     // read the data and check for same size as the working grid
-    num_read = fread(q3_array, insize, ncells, fpin);
+    num_read = (int) fread(q3_array, insize, ncells, fpin);
     fclose(fpin);
     if(num_read != NUM_CELLS)
     {
